pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vacancy-pipeline-py"
version = "0.1.0"
description = "Vacancy pipeline migrated from Node.js to Python"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "httpx>=0.28,<0.29",
  "python-dotenv>=1.0,<2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=9,<10",
  "ruff>=0.15,<0.16",
]
gmail = [
  "google-auth>=2.39,<3",
  "google-auth-oauthlib>=1.2,<2",
]

[project.scripts]
vacancy-dual-run-diff = "vacancy_pipeline_py.cli.dual_run_diff:main"
vacancy-gmail-auth = "vacancy_pipeline_py.cli.gmail_auth:main"
vacancy-orchestrate = "vacancy_pipeline_py.cli.orchestrate:main"
vacancy-send-telegram = "vacancy_pipeline_py.cli.send_telegram:main"
vacancy-sync-supabase = "vacancy_pipeline_py.cli.sync_supabase:main"
vacancy-verify = "vacancy_pipeline_py.cli.verify:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
